+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - correcting IndexError in ecs are desired tasks running probe
 - adding unit tests for are desired tasks running probe
 - updating probes ALL list to include describe functions
+- adding action to set the desired task count of an ecs service
 
 ## [0.12.0][]